+++ packages/contracts/src/core/permission/PermissionConditionUpgradeable.sol
@@ -4,18 +4,25 @@ pragma solidity ^0.8.8;
 
 import {ERC165Upgradeable} from "@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol";
 
+import {IProtocolVersion, ProtocolVersion} from "../../utils/protocol/ProtocolVersion.sol";
 import {IPermissionCondition} from "./IPermissionCondition.sol";
 
 /// @title PermissionConditionUpgradeable
 /// @author Aragon Association - 2023
 /// @notice An abstract contract for upgradeable or cloneable contracts to inherit from and to support customary permissions depending on arbitrary on-chain state.
-abstract contract PermissionConditionUpgradeable is ERC165Upgradeable, IPermissionCondition {
+/// @custom:security-contact sirt@aragon.org
+abstract contract PermissionConditionUpgradeable is
+    ERC165Upgradeable,
+    IPermissionCondition,
+    ProtocolVersion
+{
     /// @notice Checks if an interface is supported by this or its parent contract.
     /// @param _interfaceId The ID of the interface.
     /// @return Returns `true` if the interface is supported.
     function supportsInterface(bytes4 _interfaceId) public view virtual override returns (bool) {
         return
             _interfaceId == type(IPermissionCondition).interfaceId ||
+            _interfaceId == type(IProtocolVersion).interfaceId ||
             super.supportsInterface(_interfaceId);
     }
 }