+++ packages/subgraph/CHANGELOG.md
@@ -7,11 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UPCOMING]
 
+## 1.3.1
+
 ### Added
+
+- Add support for `Granted` & `Revoked` event in `PluginRepo`
+- Add one-to-many relationship from `Dao` & `PluginRepo` to `Permission`
 
 ### Changed
 
+- Renamed `fetchERC20` & `fetchWrappedERC20` to `fetchOrCreateERC20Entity` & `fetchOrCreateWrappedERC20Entity` respectively.
+- Changed type of `token` attribute of `ERC20Transfer` from `ERC20Contract` to `Token`.
+- Refactored `Permission` entity & added `pluginRepo` attribute.
+- Fixed wrong token voting member deletion, when balance & voting power was zero, but it was still delegating to another address.
+
 ### Removed
+
+- Removed `SignatureValidatorSet` event.
+- Removed `ContractPermissionId` entity.
 
 ## [1.3.0]